+++ atom/typing_utils.py
@@ -7,7 +7,19 @@
 # --------------------------------------------------------------------------------------
 import sys
 from itertools import chain
-from typing import Any, List, Tuple, TypeVar, Union, get_args, get_origin
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    List,
+    Literal,
+    Sequence,
+    Tuple,
+    TypedDict,
+    TypeVar,
+    Union,
+    get_args,
+    get_origin,
+)
 
 # In Python 3.9+, List is a _SpecialGenericAlias and does not inherit from
 # _GenericAlias which is the type of List[int] for example
@@ -22,6 +34,51 @@ if sys.version_info >= (3, 10):
     from types import UnionType
 
     UNION += (UnionType,)
+
+if TYPE_CHECKING:
+    from .atom import Atom
+
+
+class _ChangeDict(TypedDict):
+    type: Union[
+        Literal["create"],
+        Literal["update"],
+        Literal["delete"],
+        Literal["event"],
+        Literal["property"],
+        Literal["container"],
+    ]
+    name: str
+    object: "Atom"
+    value: Any
+
+
+class ChangeDict(_ChangeDict, total=False):
+    oldvalue: Any
+
+    # ContainerList specific entries, present only when type == "container"
+    operation: Union[
+        Literal["reverse"],
+        Literal["__delitem__"],
+        Literal["__iadd__"],
+        Literal["__imul__"],
+        Literal["__setitem__"],
+        Literal["append"],
+        Literal["extend"],
+        Literal["insert"],
+        Literal["pop"],
+        Literal["remove"],
+        Literal["sort"],
+    ]
+    # The following are present based on the operation value
+    olditem: Any  # operation in ("__setitem__",)
+    newitem: Any  # operation in ("__setitem__",)
+    item: Any  # operation in ("append", "insert", "pop", "remove", "__delitem__")
+    index: int  # operation in ("insert", "pop")
+    items: Sequence  # operation in ("extend", "__iadd__")
+    count: int  # operation in ("__imul__")
+    key: Any  # operation in ("sort")
+    reverse: bool  # operation in ("sort")
 
 
 def _extract_types(kind) -> Tuple[type, ...]: